Deny null codepoints when deserialising JSON
Rust and compliant JSON both permit the escaped NULL codepoint `\u0000`
to be present in UTF-8 strings.

Unfortunately, Postgres takes a different view, and will return a DB
error upon trying to insert a value containing such an escape sequence
into any JSON(B) column.

Since the legitimate use cases for a null byte in a string are
vaninshingly small, we insert JSON from requests into the DB all over
the place, and we have a fork of this library anyway, we can just
inspect the incoming JSON payload for the problematic sequence and
return an error if we find it.